+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# **1.3.0**
+
+### \*NEW\*
+#### Chat
+* Chat Mouseover tooltips
+
+<img src="https://imgur.com/BJRJBgK.png"/>
+
+### FIX
+* Fixed chat link color picker bug caused by 10.2.5 patch changes
+* Fixed and improved lfg module logic
+
+
+
 # **1.2.2**
 
 ### UPDATE